--- backend/src/main/java/org/backend/web/common/dto/EmailRequestDto.java
@@ -1,15 +1,0 @@
-package org.backend.web.common.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotEmpty;
-
-@Getter
-@Setter
-public class EmailRequestDto {
-    @Email
-    @NotEmpty(message = "이메일을 입력해 주세요")
-    private String email;
-}
