+++ backend/src/main/java/org/backend/web/common/email/dto/EmailCheckDto.java
@@ -1,0 +1,17 @@
+package org.backend.web.common.email.dto;
+
+import lombok.Data;
+
+import javax.validation.constraints.Email;
+import javax.validation.constraints.NotEmpty;
+
+@Data
+public class EmailCheckDto {
+    @Email
+    @NotEmpty(message = "이메일을 입력해 주세요")
+    private String email;
+
+    @NotEmpty(message = "인증 번호를 입력해 주세요")
+    private String authNum;
+
+}
