+++ backend/src/main/java/org/backend/web/common/email/dto/EmailRequestDto.java
@@ -1,17 +1,15 @@
-package org.backend.web.common.dto;
+package org.backend.web.common.email.dto;
 
-import lombok.Data;
+import lombok.Getter;
+import lombok.Setter;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 
-@Data
-public class EmailCheckDto {
+@Getter
+@Setter
+public class EmailRequestDto {
     @Email
     @NotEmpty(message = "이메일을 입력해 주세요")
     private String email;
-
-    @NotEmpty(message = "인증 번호를 입력해 주세요")
-    private String authNum;
-
 }
