+++ backend/src/main/java/org/backend/web/common/email/error/ExControllerAdvice.java
@@ -1,4 +1,4 @@
-package org.backend.web.error;
+package org.backend.web.common.email.error;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
